+++ src/client/src/components/DataSheets/Sections/Enterprise/Infos/EnterpriseInfos.gql.js
@@ -74,6 +74,18 @@ const effectifsMensuelsQuery = gql`
   }
 `;
 
+const effectifsPhysiqueQuery = gql`
+  query getEffectifPhysique($siren: String!) {
+    fce_Effectifs_Physiques(
+      where: { SIREN: { _eq: $siren } }
+      order_by: { DATE: desc }
+      limit: 1
+    ) {
+      EFF_TOTAL
+    }
+  }
+`;
+
 export const useEffectifsMensuels = pipe(
   (siren, limit) =>
     useQuery(effectifsMensuelsQuery, {
@@ -81,4 +93,12 @@ export const useEffectifsMensuels = pipe(
       variables: { limit, siren },
     }),
   mapQueryResult(prop("fce_entreprises_etp_effectif"))
+);
+export const useEffectifsPhysique = pipe(
+  (siren) =>
+    useQuery(effectifsPhysiqueQuery, {
+      context: { clientName: BCE_CLIENT },
+      variables: { siren },
+    }),
+  mapQueryResult(prop("fce_Effectifs_Physiques[0].EFF_TOTAL"))
 );