+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [27.56.0](https://github.com/SocialGouv/fce/compare/v27.55.0...v27.56.0) (2023-04-13)
+
+
+### Features
+
+* add effectif physique ([#255](https://github.com/SocialGouv/fce/issues/255)) ([a78782a](https://github.com/SocialGouv/fce/commit/a78782adf137e499ba93d7c65e9eb3761cd7272a))
+
+# [27.55.0](https://github.com/SocialGouv/fce/compare/v27.54.1...v27.55.0) (2023-04-13)
+
+
+### Features
+
+* integrate bce finance indicators ([#252](https://github.com/SocialGouv/fce/issues/252)) ([57d48dc](https://github.com/SocialGouv/fce/commit/57d48dc13de536f49fbc133c62b78586f00e464c))
+
 ## [27.54.1](https://github.com/SocialGouv/fce/compare/v27.54.0...v27.54.1) (2023-04-06)